signup/login finished with authContext working
